# contrato/openapi.json
{
  "components": {
    "schemas": {
      "AtsAnalysis": {
        "properties": {
          "breakdown": {
            "default": {},
            "title": "Breakdown",
            "type": "object"
          },
          "keywordsCriticasAusentes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Keywordscriticasausentes",
            "type": "array"
          },
          "keywordsEncontradas": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Keywordsencontradas",
            "type": "array"
          },
          "pontosEliminatorios": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Pontoseliminatorios",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "scoreVersao": {
            "default": 2,
            "title": "Scoreversao",
            "type": "integer"
          },
          "veredicto": {
            "title": "Veredicto",
            "type": "string"
          }
        },
        "required": [
          "score",
          "veredicto"
        ],
        "title": "AtsAnalysis",
        "type": "object"
      },
...
      "ClassifyRequest": {
        "properties": {
          "descricao": {
            "default": "",
            "title": "Descricao",
            "type": "string"
          },
          "titulo": {
            "default": "",
            "title": "Titulo",
            "type": "string"
          }
        },
        "title": "ClassifyRequest",
        "type": "object"
      },
      "ClassifyResponse": {
        "properties": {
          "categoria": {
            "title": "Categoria",
            "type": "string"
          },
          "nivel": {
            "title": "Nivel",
            "type": "string"
          }
        },
        "required": [
          "categoria",
          "nivel"
        ],
        "title": "ClassifyResponse",
        "type": "object"
      },
...
      "GenerateCvRequest": {
        "properties": {
          "contexto": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FonteContexto"
            },
            "title": "Contexto",
            "type": "array"
          },
          "keywords": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Keyword"
            },
            "title": "Keywords",
            "type": "array"
          },
          "perfilMestre": {
            "$ref": "#/components/schemas/PerfilMestre"
          },
          "vaga": {
            "$ref": "#/components/schemas/Vaga"
          }
        },
        "required": [
          "perfilMestre",
          "vaga"
        ],
        "title": "GenerateCvRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ai-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analisar-ats": {
      "post": {
        "operationId": "analisar_analisar_ats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCvRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtsAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analisar"
      }
    },
    "/classify": {
      "post": {
        "operationId": "classify_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify"
      }
    },
...
    "/copiloto/redigir-formulario": {
      "post": {
        "operationId": "copiloto_redigir_formulario_copiloto_redigir_formulario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedigirFormularioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedigirFormularioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copiloto Redigir Formulario"
      }
    },
    "/copiloto/redigir-mensagem": {
      "post": {
        "operationId": "copiloto_redigir_mensagem_copiloto_redigir_mensagem_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedigirMensagemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedigirMensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copiloto Redigir Mensagem"
      }
    },
    "/copiloto/turn": {
      "post": {
        "operationId": "copiloto_turn_copiloto_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copiloto Turn"
      }
    },
    "/copiloto/turn/stream": {
      "post": {
        "operationId": "copiloto_turn_stream_copiloto_turn_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copiloto Turn Stream"
      }
    },
    "/embeddings/consultas": {
      "post": {
        "operationId": "embeddings_consultas_embeddings_consultas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingConsultasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingConsultasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Embeddings Consultas"
      }
    },
    "/embeddings/documentos": {
      "post": {
        "operationId": "embeddings_documentos_embeddings_documentos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingDocumentosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingDocumentosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Embeddings Documentos"
      }
    },
    "/generate-cv": {
      "post": {
        "operationId": "generate_generate_cv_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCvRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateCvResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate"
      }
    },
    "/geracao/{passo}": {
      "post": {
        "operationId": "geracao_passo_geracao__passo__post",
        "parameters": [
          {
            "in": "path",
            "name": "passo",
            "required": true,
            "schema": {
              "title": "Passo",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Dados",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Geracao Passo Geracao  Passo  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geracao Passo"
      }
    },
...
    "/keywords": {
      "post": {
        "operationId": "keywords_keywords_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keywords"
      }
    },
    "/keywords/lote/interpretar": {
      "post": {
        "operationId": "interpretar_keywords_lote_keywords_lote_interpretar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretarLoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretarLoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interpretar Keywords Lote"
      }
    },
    "/keywords/lote/preparar": {
      "post": {
        "operationId": "preparar_keywords_lote_keywords_lote_preparar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepararLoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepararLoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preparar Keywords Lote"
      }
    },
    "/rag/filtrar": {
      "post": {
        "operationId": "rag_filtrar_rag_filtrar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiltroTrechosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiltroTrechosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rag Filtrar"
      }
    },
...
    "/reduzir-curriculo": {
      "post": {
        "operationId": "reduzir_reduzir_curriculo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReduzirCvRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reduzir"
      }
    },
    "/score": {
      "post": {
        "operationId": "score_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score"
      }
    }
  }
}